Further pathfinding optimization and with-fiddling.
Added multiple movement algorithms to Mall loader and a simple test mall in data/malls/simple2.bmp.
Fiddled with A* heuristic to optimize it for speed. It's not admissible anymore, meaning that generated paths might not be optimal (but are still bounded to be at most a constant factor off of optimal).